pyproject.toml
[project]
name = "sample-simple-chatapp"
version = "0.1.0"
description = ""
authors = [
    {name = "sakagami0615",email = "sakagami0615@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "streamlit (>=1.47.0,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "openai (>=1.97.0,<2.0.0)",
    "langchain (>=0.3.26,<0.4.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "duckduckgo-search (>=8.1.1,<9.0.0)",
    "wikipedia (>=1.4.0,<2.0.0)",
    "langchain-openai (>=0.3.28,<0.4.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.4"

[tool.ruff]
line-length = 119
indent-width = 4

# フォーマッター設定（blackと同じ整形ルールを適用）
[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "I",   # isort (import order)
    "UP",  # pyupgrade
    "S",   # flake8-bandit (セキュリティ)
]
ignore = [
    "E501",  # 行の長さ制限を無視（flake8と合わせるなら無効）
]

# import順の自動整形用（isort相当）
[tool.ruff.lint.isort]
combine-as-imports = true
force-single-line = false
